Rename shift_phase to phase
