+++ qiskit/pulse/instructions/__init__.py
@@ -31,5 +31,5 @@ sequence of scheduled Pulse ``Instruction`` s over many channels.
 """
 from .delay import Delay
 from .instruction import Instruction
-from .shift_phase import ShiftPhase
+from .phase import ShiftPhase
 from .snapshot import Snapshot
